+++ lib/utils/extension.dart
@@ -16,12 +16,10 @@ extension PTYExt on Pty {
     Directory dir = Directory(RuntimeEnvir.tmpPath).createTempSync();
     File('${dir.path}/shell').writeAsStringSync(function);
     // 删除这个文件夹
-    // pty.write(Uint8List.fromList(Utf8Encoder().convert('chmod +x ${dir.path}/shell\n')));
     write(Uint8List.fromList(const Utf8Encoder().convert('source ${dir.path}/shell\n')));
+    // 等待1s,source完成后删除源文件
     Future.delayed(const Duration(seconds: 1), () {
       dir.delete(recursive: true);
     });
-    // 等待1s
-    // source完成后删除源文件
   }
 }
